+++ .github/workflows/test_unitarios.yml
@@ -65,9 +65,10 @@ jobs:
                 pip install -r requirements-dev.txt
               fi
               pip install pytest pytest-cov
+              pip install pytest-xdist
           
               echo "Running tests with coverage and HTML report..."
-              pytest --cov=. --cov-report=term-missing --cov-report=html > ../../pytest_output_$service.txt
+              pytest -n auto --cov=. --cov-report=term-missing --cov-report=html > ../../pytest_output_$service.txt
           
               # Mueve el HTML coverage a una carpeta común para los artifacts
               mv htmlcov ../../coverage-reports/htmlcov-$service
